pyproject.toml
[project]
name = "voice-agent"
version = "1.0.0"
description = "Real-time conversational AI voice agent using LangChain, Whisper, and Piper TTS"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Voice Agent Team"}
]
keywords = ["voice-agent", "langchain", "whisper", "tts", "ai", "conversational-ai"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "langchain>=0.3.0",
    "langchain-ollama>=0.2.0",
    "langchain-core>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain-huggingface>=0.0.3",
    "docling>=1.0.0",
    "sentence-transformers>=2.2.2",
    "fpdf2>=2.7.0",
    "assemblyai>=0.23.0",
    "duckduckgo-search>=5.0.0",
    "piper-tts>=1.2.0",
    "onnxruntime>=1.16.0",
    "numpy>=1.24.0",
    "python-dotenv>=1.0.0",
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "websockets>=12.0",
    "sounddevice>=0.5.0",
    "aiohttp>=3.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "black>=24.0.0",
    "ruff>=0.3.0",
]

[project.scripts]
voice-agent = "voice_agent.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["voice_agent"]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
